requirements.txt
# ======================================
# 🌐 Django Core
# ======================================
Django==5.2.8
djangorestframework==3.15.2
asgiref==3.8.1
sqlparse==0.5.1
tzdata==2025.1

# ======================================
# ⚡ Real-time & WebSocket
# ======================================
channels==4.1.0
daphne==4.1.2
twisted==24.3.0
autobahn==23.6.2
async-timeout==4.0.3

# ======================================
# 🔌 Raspberry / Arduino Communication
# ======================================
pyserial==3.5
requests==2.32.3

# ======================================
# 📸 Camera & Image Processing
# ======================================
opencv-python==4.10.0.84
Pillow==10.4.0
numpy==1.26.4

# ======================================
# 🧱 Server Utilities
# ======================================
whitenoise==6.7.0          # 정적파일 서빙
gunicorn==22.0.0           # Linux 서버용 WSGI
uvicorn==0.30.1            # ASGI 서버 (테스트용)
psutil==5.9.8              # 리소스 모니터링 (선택)
python-dotenv==1.0.1       # 환경변수 관리

# ======================================
# 🧰 Developer Tools
# ======================================
black==24.8.0
flake8==7.1.1
isort==5.13.2
